feat: implement new useRuns hook and RunsTable component
